improve progress bar logging
